pyproject.toml
[project]
name = "surek"
version = "2.0.0"
description = "Docker Compose orchestration tool for self-hosted services"
readme = "README.md"
license = { text = "GPL-3.0-or-later" }
requires-python = ">=3.12"
authors = [
    { name = "Oleg Wock" }
]
keywords = ["docker", "compose", "orchestration", "self-hosted", "caddy", "reverse-proxy"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]

dependencies = [
    "pydantic>=2.0",
    "textual>=0.50",
    "typer>=0.12",
    "rich>=13.0",
    "docker>=7.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "boto3>=1.34",
    "bcrypt>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.8",
    "ruff>=0.3",
    "types-pyyaml",
    "types-boto3",
    "boto3-stubs[s3]",
]

[project.scripts]
surek = "surek.cli.main:app"

[project.urls]
Homepage = "https://github.com/OlegWock/surek"
Repository = "https://github.com/OlegWock/surek"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/surek"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
